requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
telethon==1.32.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2 